* libmisc/console.c, libmisc/hushed.c, libmisc/yesno.c, 	libmisc/loginprompt.c, libmisc/ttytype.c, libmisc/tz.c, 	src/login_nopam.c, src/chpasswd.c, src/chgpasswd.c, lib/port.c: 	The size argument of fgets is an int, not a size_t. 	* libmisc/loginprompt.c: Ignore the return value from signal() 	when the signal handlers are restored. 	* src/chpasswd.c: Cast the return value of time() to a long 	integer. 	* src/chpasswd.c: Use the SCALE macro instead of (24L * 3600L) 	for the values to be set in /etc/shadow.

+++ lib/port.c
@@ -157,7 +157,7 @@ static struct port *getportent (void)
 	 * is a '\n'.  Lines which begin with '#' are all ignored.
 	 */
 
-	if (fgets (buf, sizeof buf, ports) == 0) {
+	if (fgets (buf, (int) sizeof buf, ports) == 0) {
 		errno = saveerr;
 		return 0;
 	}
