+++ libmisc/ttytype.c
@@ -62,7 +62,7 @@ void ttytype (const char *line)
 		perror (typefile);
 		return;
 	}
-	while (fgets (buf, sizeof buf, fp) == buf) {
+	while (fgets (buf, (int) sizeof buf, fp) == buf) {
 		if (buf[0] == '#') {
 			continue;
 		}
